Sec 4.22 - 4.25
TakableHeart by C.Nilsson from OpenGameArt.org.
Sec 4.23: I made the Heart collectable but did not give it the "Shiny"
material texture as the tutorial instructs.

Sec 4.24: I wrote the RandomSprite code but did not apply it to any
scenery sprites. I prefer permanent scenery, but the code is there if we
want to change it in the future.

Sec 4.25: Updated the RandomSprite code for random sorting order, but
still did not apply it to any scenery sprites.