Sec 4.27-4.28
Have not yet split the background into separate layers. Will need to ask
Dan if he has a program which can accomplish this; or download Photoshop
or Gimp. Currently, there is only on layer which I could not change to
type "Texture" so it is still type "Sprite 2D and UI". Code is
completed, and will not need to be changed. However, in Unity the public
variables under FollowCamera need to add layers if we choose to do so
later.